Support newer Linux versions.

+++ include/click/ino.hh
@@ -50,8 +50,8 @@ class ClickIno { public:
 
     // readdir handles '..' (f_pos 0) and '.' (f_pos 1).
     // It returns the number of things stored.
-    typedef bool (*filldir_t)(const char *name, int name_len, ino_t ino, int dirtype, uint32_t f_pos, void *user_data);
-    int readdir(ino_t dir, uint32_t &f_pos, filldir_t fd, void *user_data);
+    typedef bool (*filldir_t)(const char *name, int name_len, ino_t ino, int dirtype, loff_t f_pos, void *user_data);
+    int readdir(ino_t dir, loff_t &f_pos, filldir_t fd, void *user_data);
 
 #if INO_DEBUG
     String info() const;
